[#39] :recycle: refactor: redirect 함수 모듈화, sns login 처리 hook sns 컴포넌트 내부로 이사
